Introduce transparent batching of JSON-RPC requests via web3 proxying.
The solution is based on ES6 proxying of the web3.eth object. We
integrate with the DataLoader library which orchestrates the batching
of individual key retrievals. For calls elegible for batching, the proxy
captures the call in a command object and returns a the DataLoader
Promise, thus preserving the contract of web3.eth. When a batch is
closed, all direct calls are converted to their batched version and
the batch is executed. Callback results are then spread back to their
original promises.

Closes #38.