Added requested changes
Further documentation including JSDoc and putting the solidity contract in the test as a comment. Also minor rework of value function to use a temporary base to support mulitple value calls (and added a test for mulitple value calls). Using string unions for storageType and base and previousType begin their lives as null and not '' now